+++ generic_daw_core/src/time/beat_time.rs
@@ -172,36 +172,36 @@ impl SubAssign for BeatTime {
 	}
 }
 
-impl Mul<f32> for BeatTime {
+impl Mul<f64> for BeatTime {
 	type Output = Self;
 
-	fn mul(self, rhs: f32) -> Self::Output {
+	fn mul(self, rhs: f64) -> Self::Output {
 		Self(self.0 * rhs)
 	}
 }
 
-impl MulAssign<f32> for BeatTime {
-	fn mul_assign(&mut self, rhs: f32) {
+impl MulAssign<f64> for BeatTime {
+	fn mul_assign(&mut self, rhs: f64) {
 		*self = *self * rhs;
 	}
 }
 
-impl Div<f32> for BeatTime {
+impl Div<f64> for BeatTime {
 	type Output = Self;
 
-	fn div(self, rhs: f32) -> Self::Output {
+	fn div(self, rhs: f64) -> Self::Output {
 		Self(self.0 / rhs)
 	}
 }
 
-impl DivAssign<f32> for BeatTime {
-	fn div_assign(&mut self, rhs: f32) {
+impl DivAssign<f64> for BeatTime {
+	fn div_assign(&mut self, rhs: f64) {
 		*self = *self / rhs;
 	}
 }
 
 impl Div for BeatTime {
-	type Output = f32;
+	type Output = f64;
 
 	fn div(self, rhs: Self) -> Self::Output {
 		self.0 / rhs.0
