+++ generic_daw_core/src/time/fixed_point.rs
@@ -127,38 +127,38 @@ impl SubAssign for FixedPoint {
 	}
 }
 
-impl Mul<f32> for FixedPoint {
+impl Mul<f64> for FixedPoint {
 	type Output = Self;
 
-	fn mul(self, rhs: f32) -> Self::Output {
-		Self((self.0 as f64 * f64::from(rhs)) as u64)
+	fn mul(self, rhs: f64) -> Self::Output {
+		Self((self.0 as f64 * rhs) as u64)
 	}
 }
 
-impl MulAssign<f32> for FixedPoint {
-	fn mul_assign(&mut self, rhs: f32) {
+impl MulAssign<f64> for FixedPoint {
+	fn mul_assign(&mut self, rhs: f64) {
 		*self = *self * rhs;
 	}
 }
 
-impl Div<f32> for FixedPoint {
+impl Div<f64> for FixedPoint {
 	type Output = Self;
 
-	fn div(self, rhs: f32) -> Self::Output {
-		Self((self.0 as f64 / f64::from(rhs)) as u64)
+	fn div(self, rhs: f64) -> Self::Output {
+		Self((self.0 as f64 / rhs) as u64)
 	}
 }
 
-impl DivAssign<f32> for FixedPoint {
-	fn div_assign(&mut self, rhs: f32) {
+impl DivAssign<f64> for FixedPoint {
+	fn div_assign(&mut self, rhs: f64) {
 		*self = *self / rhs;
 	}
 }
 
 impl Div for FixedPoint {
-	type Output = f32;
+	type Output = f64;
 
 	fn div(self, rhs: Self) -> Self::Output {
-		(self.0 as f64 / rhs.0 as f64) as f32
+		self.0 as f64 / rhs.0 as f64
 	}
 }
